# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlyRank Auth API · Login & Protect",
    "version": "1.0.0",
    "description": "A secure API utilizing Supabase Auth as Identity Provider, JWT verification middleware, and Swagger UI interactive documentation with Bearer authentication."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Supabase JWT access token obtained from /auth/login"
      }
    },
    "schemas": {
      "AuthCredentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "ref_9a8b7c6d5e4f3a2b1c0d"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid or expired token"
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "test@example.com"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-09T14:30:00.000Z"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create a new user account",
        "description": "Registers a new user in Supabase Auth. Validates email and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user and return JWT access token",
        "description": "Authenticates user credentials with Supabase and returns an access token and refresh token. Protected with brute-force rate limiting.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOi..."
                    },
                    "refresh_token": {
                      "type": "string",
                      "example": "ref_9a8b7c..."
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserProfile"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid login credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts (Rate limited)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Exchange refresh token for fresh access token",
        "description": "Allows clients to acquire a new access token when the current short-lived JWT expires without requiring the user to re-enter credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing refresh token"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign out and terminate session",
        "description": "Invalidates the current session with Supabase. Requires a valid Bearer token.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully logged out (No Content)"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Read public data",
        "description": "Open endpoint accessible to everyone without authentication.",
        "responses": {
          "200": {
            "description": "Public welcome message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Welcome stranger! This info is public."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get authenticated user profile",
        "description": "Extracts and verifies the Bearer JWT token with Supabase and returns safe user metadata.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing, malformed, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get authenticated user dashboard",
        "description": "Second protected endpoint demonstrating auth middleware reusability.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User dashboard information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Welcome to your dashboard, test@example.com!"
                    },
                    "user_id": {
                      "type": "string",
                      "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    },
                    "account_created": {
                      "type": "string",
                      "example": "2026-09-09T14:30:00.000Z"
                    },
                    "role": {
                      "type": "string",
                      "example": "authenticated"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/protected/admin": {
      "get": {
        "tags": ["Protected"],
        "summary": "Access privileged admin route",
        "description": "Demonstrates 403 Forbidden vs 401 Unauthorized. User must be authenticated AND have admin privileges.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin console access granted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "admin_user": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid"
          },
          "403": {
            "description": "Forbidden - Authenticated user lacks admin permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}